Update draft-bill.md
I updated the numbering and most cross-references to reflect the future location of the legislation in Minn. Stat. § 80A.46.  I edited the language and grammatical structure through paragraph (26)(A) "Qualifications" to conform to the language and grammatical structure found in Minn. Stat. § 80A.46; for example, see paragraph (14).  This should be continued at later time through paragraphs (26)(B)-(M).